+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "payload-swagger",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "description": "Swagger plugin for payload cms",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
