+++ src/index.ts
@@ -1,72 +1,6 @@
-import type { Express } from 'express';
-import { Payload } from 'payload';
-import { Config, SanitizedConfig } from 'payload/config';
-import swaggerUi from 'swagger-ui-express';
-import { createDocument } from './open-api';
-import { Options } from './types';
-import { serveFile } from './utils/serve-file';
+export type { Options } from './types';
 
 export { createDocument } from './open-api';
+export { swagger } from './plugin';
 
-/**
- * Add swagger routes to a payload server
- */
-const loadSwagger = async (app: Express, config: SanitizedConfig, options?: Options) => {
-  const document = await createDocument(config, options);
-
-  app.use('/api-docs/specs', (req, res) => res.json(document));
-  if (document.info.license?.url) {
-    app.get('/api-docs/license', serveFile('LICENSE'));
-  }
-
-  app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(undefined, { swaggerUrl: '/api-docs/specs' }));
-};
-
-/**
- * Payload swagger plugin
- */
-export const swagger =
-  (options?: Options) =>
-  (config: Config): Config => ({
-    ...config,
-    admin: {
-      ...config.admin,
-      webpack: webpackConfig => {
-        const modifiedConfig = {
-          ...webpackConfig,
-          resolve: {
-            ...webpackConfig.resolve,
-            fallback: {
-              fs: false,
-              async_hooks: false,
-              net: false,
-              stream: require.resolve('stream-browserify'),
-              url: require.resolve('url'),
-              util: false,
-              zlib: false,
-              ...webpackConfig.resolve?.fallback,
-            },
-          },
-        } as any;
-        if (config.admin?.webpack) {
-          return config.admin.webpack(modifiedConfig);
-        }
-        return modifiedConfig;
-      },
-    },
-    onInit: async (payload: Payload) => {
-      if (payload.express) {
-        await loadSwagger(payload.express, payload.config, options);
-      } else {
-        payload.logger.warn('Unable to load swagger: express not available');
-      }
-
-      if (config.onInit) {
-        return config.onInit(payload);
-      }
-    },
-  });
-
-export default loadSwagger;
-
-export { Options };
+export { loadSwagger as default } from './swagger';
